Added 2 swr hooks: useProject and useUsage & refactored code
