+++ components/app/custom-domain/index.tsx
@@ -12,10 +12,12 @@ import {
 import { useEditDomainModal } from "./edit-domain-modal";
 import { DomainVerificationStatusProps } from "@/lib/types";
 import DomainConfiguration from "./domain-configuration";
+import useProject from "@/lib/swr/use-project";
 
-export default function CustomDomain({ domain }: { domain: string }) {
+export default function CustomDomain() {
   const router = useRouter();
   const { slug } = router.query as { slug: string };
+  const { project: { domain } = {} } = useProject();
 
   const { data, isValidating } = useSWR<{
     status: DomainVerificationStatusProps;
