+++ components/app/link-card.tsx
@@ -9,7 +9,7 @@ import { LinkProps } from "@/lib/types";
 import Tooltip, { TooltipContent } from "@/components/shared/tooltip";
 import useProject from "@/lib/swr/use-project";
 import useUsage from "@/lib/swr/use-usage";
-import { useAddEditLinkModal } from "./add-link-modal";
+import { useAddEditLinkModal } from "./add-edit-link-modal";
 import { useDeleteLinkModal } from "./delete-link-modal";
 
 export default function LinkCard({ props }: { props: LinkProps }) {
